refactor validation into runresult, not done yet
